Update CreateSVG()
I have changed how the SVG file is written. I think this new format is easier to work with once I open it in Adobe Illustrator. Similar to how the PNG file is aligned with original image (in publication), I can open the output SVG file and overlay the original image in that file. Since the artboard is of the same aspect ratio of the original image, I simply resize the original image to the artboard size to achieve alignment. Not sure if it helps.